pyproject.toml
[tool.poetry]
name = "arguelauncher"
version = "1.0.0"
description = "Evaluation framework for argumentation machines"
authors = ["Mirko Lenz <info@mirko-lenz.de>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
typer = "^0.7.0"
nlp-service = "^1.2.2"
arg-services = "^1.1.0"
arguebuf = { version = "^2.0.0b15", allow-prereleases = true }
hydra-core = "^1.3.1"
hydra-joblib-launcher = "^1.2.0"
rich = "^13.3.1"
typing-extensions = "^4.5.0"

[build-system]
requires = ["poetry"]
build-backend = "poetry.masonry.api"
